ai/requirements.txt
fastapi
uvicorn
pillow
numpy
torch
detectron2
